ARM: mach-imx6q: add ksz9131rn_phy_fixup
The MAC of the i.MX6 SoC is compliant with RGMII v1.3. The KSZ9131 PHY
is like KSZ9031 adhering to RGMII v2.0 specification. This means the
MAC should provide a delay to the TXC line. Because the i.MX6 MAC does
not provide this delay this has to be done in the PHY.

This patch adds by default ~1.4ns delay to the TXC line. This should
be good for all boards that have all RGMII signals routed with the
same length.

The KSZ9131 has relatively high tolerances on skew registers from
MMD 2.4 to MMD 2.8. Therefore the new DLL-based delay of 2ns is used
and then as little as possibly subtracted from that so we get more
accurate delay. This is actually needed because the i.MX6 SoC has
an asyn skew on TXC from -100ps to 900ps.

Related-to: ELB-1299
Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>