ARM: dts: imx6ull-colibri: disable 1.8V modes for Evaluation Board
Move the SD-card properties to the module level device tree since this
is the default Colibri SD-card interface.

However, disable 1.8V for the Colibri Evaluation Board since this
carrier board has 3.3V pull-ups on

The Evaluation Carrier Board has 3.3V pull-ups on the carrier board
which seems to make 1.8V modes not usable. Using a UHS-I capable SD-card
leads to the following error:
[  164.058099] mmc0: error -110 whilst initialising SD card

Using the no-1-8-v property makes this restriction obvious and fixes the
issue.

Related-to: TEI-472
Fixes: 20353143a27f4d ("ARM: dts: imx6ull-colibri: enable UHS-I for USDHC1")
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>