ARM64: dts: apalis-imx8qm: default to a 3.3v rail for the ext rgmii signals
The pins which can be used for an external ethernet PHY are powered by
a sepearte LDO. Currently in the device tree this rail is set to 1.8V,
the default apalis functionality of these pins requires 3.3V so change
to that.

If used with an external PHY in addition to add the PHY to the device also
the rail's voltage needs to be adapted.

Related-to: ELB-2899
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>