ARM: dts: colibri-imx7: fix wake-up with gpio key
The pin GPIO1_IO01 externally pulls down, it is required to sequentially
connect this pin (signal WAKE_MICO#) to +3v3 and then disconnect it to
trigger a wakeup interrupt.
Adding the flag GPIO_PULL_DOWN allows the system to be woken up just
connecting the pin GPIO1_IO01 to +3v3.

Related-to: ELB-2901
Signed-off-by: Oleksandr Suvorov <oleksandr.suvorov@toradex.com>