Revert "arm64: dts: imx8mm-verdin: add 32.768 khz clock to pmic"
Our hardware routes the RTC clock via PMIC to the i.MX 8M Mini SoC.
However, the PMIC is not aware of this and just turns the clock off
during boot thinking it is unused. Solving this properly proves
non-trivial but with the upcoming re-designed modules using another PMIC
this anyway all changes again. So just reverting this for now.

This reverts commit dd1060116f5fbeb63465e97f3cb8022c9b4bff4c.

Signed-off-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>