fec_main.c: disable/enable phy irq on suspend/resume
Commit 07b0928918c6 ("net: phy: enable interrupts when PHY is attached
already" dropped disable/enable the phy interrupt on stop/start.

In the suspend/resume case, if a phy regulator is configured that
regulator is switched off and thus the phy interrupt is constantely
asserted leading to unserviceable interrupts after suspend.

Disable/enable the irq in the MAC's suspend/resume functions if
a phy regulator is configured.

Related-to: ELB-2757
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>