pci: pci-imx6: fix kernel hanging on imx6q
The resume from a suspend state on SoC i.MX6Q leads to a
kernel hanging.
Adding a 10ms delay after clearing the bit TEST_POWERDOWN
of IOMUXC_GPR1 regiater fixes this issue.

There is not enough information about the testing "powerdown"
mode on i.MX6Q, but we can do some assumption.
According to "49.6.1.2 Power-On Reset" section of the
"i.MX6D/6Q APRM", the time between bringing up the PHY power
supplies and the reference clock is up and stable is
(< 10ms) + (> 10mks) ~= 10ms. So using this value seems reasonable.

Related-to: ELB-3012
Signed-off-by: Oleksandr Suvorov <oleksandr.suvorov@toradex.com>