dts: colibri-imx6ull: reenable pwm4 and i2c2
If PWM_A is not used for the backlight it might be used as regular PWM
through the sysfs, so keep enabled.

If i2c2 is disabled this might change the bus numbering of the I2C
busses. Keep the bus enabled, even if we do not enable the touch
controller attached to it.

Extends commit 1c530823b1e9 ("dts: colibri-imx6ull: added overlays split").

Related-to: ELB-4065
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>